requirements.txt
python-docx>=0.8.11
Flask>=3.0.0
Werkzeug>=3.0.0
openai>=1.0.0
